Add messages for TODO onto schematic
Things are in pretty rough shape.  I had bad habits about saving revisions
and whatnot, so now this is where we are.

Now, so it is easier for multiple people to work on it, I've marked a
bunch things that need doing.

Also on the list:  Box/enclosure design, and a possible backplane.